Filter history items based on status
This commit updates the `HistoryScreen` to filter the displayed history items based on a selected status filter.

Key changes:
- In `HistoryScreen.kt`:
    - The list of history items is now filtered based on the `selectedFilter` state.
    - `LazyColumn` now uses the `filteredItems` list.
    - Added minor padding adjustments to the `LazyColumn`.

+++ app/src/main/java/com/example/doctorappoint/ui/service/BookingScreen.kt
@@ -51,7 +51,7 @@ fun BookingScreen(
     navController: NavHostController,
     modifier: Modifier = Modifier,
     departmentId: Int,
-    price: Double
+    price: Int
 ){
     val departmentViewModel: DepartmentViewModel = viewModel()
     val departmentsState by departmentViewModel.departments.collectAsState()
